Rewrote Apollo integration
