+++ pages/user/[username].tsx
@@ -3,6 +3,7 @@ import { useRouter } from 'next/router'
 import PropTypes from 'prop-types'
 import gql from 'graphql-tag';
 import { useQuery } from '@apollo/client';
+import { withApollo } from '../../lib/apollo';
 
 import CollectionComponent from '../../src/components/collection/Collection';
 import collectionMockData from '../../data/mock-data-1.json';
@@ -11,7 +12,7 @@ interface Props {
 
 }
 
-export default function UserPage({}: Props): ReactElement {
+function UserPage({}: Props): ReactElement {
   const router = useRouter();
   const { username } = router.query;
 
@@ -27,21 +28,22 @@ export default function UserPage({}: Props): ReactElement {
     }
   `;
 
-  let collection: any;
   let loading, error, data;
   if(username === 'mock') {
-    collection = collectionMockData;
+    data = {collection: collectionMockData};
     loading = error = false;
   }
   else {
     ({ loading, error, data } = useQuery(GET_COLLECTION_QUERY));
-    collection = data.collection
   }
   if (loading) return <p>Loading...</p>;
-  if (error) return <p>Error: {error}</p>;
+  if (error) return <p>Error</p>;
   return (
     <div>
-      <CollectionComponent collection={collection}></CollectionComponent>
+      <CollectionComponent collection={data.collection}></CollectionComponent>
     </div>
   );
 }
+
+// export default withApollo(UserPage);
+export default UserPage;
